+++ GUI-dev/src/app/@theme/components/index.ts
@@ -3,3 +3,4 @@ export * from './footer/footer.component';
 export * from './search-input/search-input.component';
 export * from './theme-settings/theme-settings.component';
 export * from './theme-switcher/theme-switcher.component';
+export * from './tac/tac.component';