+++ GUI-dev/src/app/pages/pages-menu.ts
@@ -3,11 +3,10 @@ import { NbMenuItem } from '@nebular/theme';
 export const MENU_ITEMS: NbMenuItem[] = [
   {
     title: 'Dashboard',
-    icon: 'nb-home',
-    link: '/pages/dashboard',
-    home: true,
+    icon: 'nb-bar-chart',
+    link: '/dashboard',
   },
-  {
+  /*{
     title: 'FEATURES',
     group: true,
   },
@@ -32,5 +31,5 @@ export const MENU_ITEMS: NbMenuItem[] = [
         link: '/auth/reset-password',
       },
     ],
-  },
+  },*/
 ];