+++ GUI-dev/src/app/pages/pages-routing.module.ts
@@ -3,6 +3,9 @@ import { NgModule } from '@angular/core';
 
 import { PagesComponent } from './pages.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
+import { HomeComponent } from './home/home.component';
+import {AuthGuard} from '../auth-guard.service';
+
 
 const routes: Routes = [{
   path: '',
@@ -11,10 +14,15 @@ const routes: Routes = [{
     {
       path: 'dashboard',
       component: DashboardComponent,
+      canActivate: [AuthGuard],
+    },
+    {
+      path: 'home',
+      component: HomeComponent,
     },
     {
       path: '',
-      redirectTo: 'dashboard',
+      redirectTo: 'home',
       pathMatch: 'full',
     },
   ],