+++ GUI-dev/src/app/role.provider.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs/Observable';
+import { map } from 'rxjs/operators/map';
+
+import { NbAuthService, NbAuthJWTToken } from '@nebular/auth';
+import { NbRoleProvider } from '@nebular/security';
+
+@Injectable()
+export class RoleProvider implements NbRoleProvider {
+
+  constructor(private authService: NbAuthService) {
+  }
+
+  getRole(): Observable<string> {
+    return this.authService.onTokenChange()
+      .pipe(
+        map((token: NbAuthJWTToken) => {
+          // temporary using first role
+          // todo: implement proper role array utilization
+          return token.isValid() ? token.getPayload()['roles'][0] : 'guest';
+        }),
+      );
+  }
+}